fix: refactor exception imports and enhance error handling in sample_caller
