+++ juriscraper/lib/exceptions.py
@@ -54,3 +54,35 @@ class InvalidDocumentError(Exception):
     or our `expected_content_type` controls. Proper place to raise this
     would be on `Site.cleanup_content`
     """
+
+
+class BadContentError(JuriscraperException):
+    """Parent class for errors raised when downloading binary content"""
+
+
+class UnexpectedContentTypeError(BadContentError):
+    """Occurs when the content received from the server has
+    a different content type than the ones listed on
+    site.expected_content_types
+    """
+
+    def __init__(self, message: str, fingerprint: str = None):
+        self.fingerprint = fingerprint
+        log_message = message
+        if fingerprint:
+            log_message += f" | Fingerprint: {fingerprint}"
+        self.logger.error(log_message)
+
+
+class NoDownloadUrlError(BadContentError):
+    """Occurs when a DeferredList fetcher fails."""
+
+    def __init__(self, message: str):
+        self.logger.error(message)
+
+
+class EmptyFileError(BadContentError):
+    """Occurs when the content of the response has lenght 0"""
+
+    def __init__(self, message: str):
+        self.logger.error(message)
